+++ great_expectations/data_context/data_context/abstract_data_context.py
@@ -50,7 +50,6 @@ from great_expectations.core.config_provider import (
 )
 from great_expectations.core.expectation_validation_result import get_metric_kwargs_id
 from great_expectations.core.id_dict import BatchKwargs
-from great_expectations.core.metric import ValidationMetricIdentifier
 from great_expectations.core.run_identifier import RunIdentifier
 from great_expectations.core.serializer import (
     AbstractConfigSerializer,
@@ -91,6 +90,7 @@ from great_expectations.data_context.types.refs import (
 from great_expectations.data_context.types.resource_identifiers import (
     ConfigurationIdentifier,
     ExpectationSuiteIdentifier,
+    ValidationMetricIdentifier,
     ValidationResultIdentifier,
 )
 from great_expectations.data_context.util import (
