+++ tests/actions/test_store_metric_action.py
@@ -6,10 +6,10 @@ from great_expectations.core.expectation_validation_result import (
     ExpectationSuiteValidationResult,
     ExpectationValidationResult,
 )
-from great_expectations.core.metric import ValidationMetricIdentifier
 from great_expectations.core.run_identifier import RunIdentifier
 from great_expectations.data_context.types.resource_identifiers import (
     ExpectationSuiteIdentifier,
+    ValidationMetricIdentifier,
     ValidationResultIdentifier,
 )
