+++ tests/data_context/store/test_evaluation_parameter_store.py
@@ -11,7 +11,6 @@ from great_expectations.core.expectation_validation_result import (
     ExpectationSuiteValidationResult,
     ExpectationValidationResult,
 )
-from great_expectations.core.metric import ValidationMetricIdentifier
 from great_expectations.core.run_identifier import RunIdentifier
 from great_expectations.data_context.data_context import DataContext
 from great_expectations.data_context.store import (
@@ -19,6 +18,9 @@ from great_expectations.data_context.store import (
     TupleAzureBlobStoreBackend,
     TupleGCSStoreBackend,
     TupleS3StoreBackend,
+)
+from great_expectations.data_context.types.resource_identifiers import (
+    ValidationMetricIdentifier,
 )
 from great_expectations.data_context.util import instantiate_class_from_config
 from tests.core.usage_statistics.util import (
