+++ tests/data_context/store/test_evaluation_parameter_store_pre_v013.py
@@ -10,8 +10,10 @@ from great_expectations.core.expectation_validation_result import (
     ExpectationSuiteValidationResult,
     ExpectationValidationResult,
 )
-from great_expectations.core.metric import ValidationMetricIdentifier
 from great_expectations.core.run_identifier import RunIdentifier
+from great_expectations.data_context.types.resource_identifiers import (
+    ValidationMetricIdentifier,
+)
 from great_expectations.data_context.util import instantiate_class_from_config
